requirements.txt
crewai==0.130.0
langchain==0.3.25
langchain-google-genai==2.1.5
httpx==0.28.1
python-decouple==3.8
asyncio==3.4.3
psycopg2-binary==2.9.10
asyncpg==0.30.0
requests==2.32.4
sseclient-py==1.8.0
google-ai-generativelanguage==0.6.18
email-validator==2.2.0
pydantic==2.11.7
mcp==1.9.4
websocket-client==1.8.0
fastmcp==2.8.1
langchain-groq==0.3.2
boto3==1.33.6
python-dotenv==1.1.0
langchain-core==0.3.65
langchain-community==0.3.25
crewai-tools==0.47.1
tenacity==9.1.2
openai==1.88.0
psycopg2==2.9.10
flask==3.0.3
gunicorn==21.2.0